+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Total points with tactics in fantasy race detail
 ### Changed
 - Center the nav-tabs
+- Simplify the Fantasy League stats page
 
 ## [2.16.11] - 2025-03-16
 ### Fixed